requirements.txt
asgiref==3.2.3
Click==7.0
Django==3.0.3
django-picklefield==2.1.1
gunicorn==20.0.4
livereload==2.6.1
Markdown==3.2.1
MarkupSafe==1.1.1
mkdocs==1.0.4
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
PyYAML==5.3
redis==3.4.1
six==1.14.0
sqlparse==0.3.0
tornado==6.0.3

djangorestframework~=3.12.4